demo_servers/remap/requirements.txt
django==1.11
djangorestframework==3.6.3
markdown==2.6.8
django-filter==1.0.4
django-bootstrap-form==3.2.1
djangorestframework-yaml==1.0.3
djangorestframework-jsonp==1.0.2
django_compressor==2.1.1
django-debug-toolbar==1.8
python-memcached==1.58
coreapi
django-rest-swagger